requirements.txt
Django==1.11.23
Pillow==6.2.0
djangorestframework